fix(lsp): handle nil response case in WriteResponse function refactor(vet): add PreRunE to cobra command in seltabls cmd fix(vet): update error message when no files provided refactor(vet): use diagnostics from vetFile response
