+++ tools/seltabls/pkg/server/handle.go
@@ -166,11 +166,7 @@ func HandleMessage(
 					return nil, fmt.Errorf("failed to read file: %w", err)
 				}
 				state.Documents[request.Params.TextDocument.URI] = string(read)
-				response := lsp.Response{
-					RPC: lsp.RPCVersion,
-				}
-
-				return response, nil
+				return nil, nil
 			case methods.MethodNotificationTextDocumentDidClose:
 				var request lsp.DidCloseTextDocumentParamsNotification
 				if err = json.Unmarshal([]byte(msg.Content), &request); err != nil {
